create hook to separte logic state, validations and more from ProfileForm
